backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
python-dotenv
pyJWT
pytest
httpx
pytest-asyncio
